requirements.txt
backcall==0.1.0
bleach==3.0.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.5
decorator==4.0.11
defusedxml==0.5.0
entrypoints==0.2.3
future==0.17.1
fuzzywuzzy==0.17.0
GetOldTweets3==0.0.9
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.5
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mpmath==1.0.0
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.2
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.2.0
praw==6.0.0
prawcore==1.0.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
Pygments==2.3.0
pyparsing==2.3.0
pyquery==1.4.0
pytesseract==0.2.4
python-dateutil==2.7.5
python-Levenshtein==0.12.0
python-twitter==3.5
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.19.1
requests-oauthlib==1.0.0
scipy==1.1.0
Send2Trash==1.5.0
six==1.10.0
sympy==1.3
terminado==0.8.1
tesseract==0.1.3
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
update-checker==0.16
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
